requirements.txt
# Backend
fastapi==0.109.0
uvicorn==0.27.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
neo4j==5.15.0
SQLAlchemy==2.0.25
openai>=1.0.0
uuid==1.30
aiosqlite==0.19.0

# Frontend
streamlit==1.30.0
requests==2.31.0
matplotlib